[Feature] Create task form css modal effect
