+++ src/dom-manipulation.js
@@ -1,6 +1,22 @@
 import { collector, addItems } from "./container";
 import { saveTask, retrieveTasks } from "./storage";
 
+const showTaskForm = () => {
+  const showBut = document.querySelector('.add-task');
+  const modal = document.querySelector('.modal');
+  showBut.onclick = () => {
+    modal.style.display = "block";
+  }
+  document.querySelector('.close').onclick = () => {
+    modal.style.display = "none";
+  }
+  window.onclick = (e) => {
+    if (e.target === modal) {
+      modal.style.display = "none";
+    }
+  }
+}
+
 let projects = [];
 const appendProjects = stored => {
   const select = document.querySelector("#project");
@@ -24,6 +40,7 @@ const appendTask = (elem, ind) => {
   const newElem = document.createElement("div");
   newElem.classList.add("list-item");
   newElem.innerHTML = `
+  <span class="list-item-before"></span>
   <div class="task-item" data-index=${ind}>
       <input type="checkbox">
       <h2>${elem.title}</h2>
@@ -47,7 +64,7 @@ const displayTasks = () => {
 };
 
 const addTasks = () => {
-  var form = document.querySelector("form");
+  var form = document.querySelector(".form");
   form.addEventListener("submit", e => {
     e.preventDefault();
     const arr = ["", "", "", "", "", "", ""];
@@ -91,16 +108,17 @@ const completeTask = () => {
     checkboxes.forEach(box => {
     
      box.onclick = () => {
-               
+       const parent = box.parentNode
       if (box.checked == true) {
-        const parent = box.parentNode
         const ind = parent.dataset.index;
         parent.parentNode.style.backgroundColor = "rgba(252, 87, 101, 0.29)"
         parent.parentNode.style.borderTop = "2px solid green"
         parent.parentNode.style.color = "gray"
+        document.querySelectorAll('.list-item-before')[ind].style.width = "100%"
         // parent.parentNode.style.setProperty('--width', `${100}%`)
-          console.log(ind);
-        editTask(2);
+        console.log(document.styleSheets);
+        
+        // editTask(2);
       } else {
         parent.parentNode.style.borderTop = "2px solid black"
         parent.parentNode.style.backgroundColor = "rgba(252, 87, 101, 0.829)"
@@ -113,65 +131,6 @@ const completeTask = () => {
   });
 };
 
-const editTask = (index) => {
-
-    const task = retrieveTasks().array[index]
-    console.log(task)
-    let formBase = ` <form action="" method="post" class="edit-form">
-          <div class="choose-project">
-            <label for="project">Choose Existing Project</label>
-            <select id="project" name="project">
-              <option value="">${task.project}</option>
-            </select>
-            <span>or</span>
-            <div>
-              <label for="newProj">Add new Project</label>
-              <input type="text" name="newProj" id="newProj">
-            </div>
-          </div>
-          <div>
-            <label for="title">Task Title</label>
-            <input type="text" name="title" id="title" value="${task.title}">
-          </div>
-          <div>
-            <label for="title">Task Description</label>
-            <input type="text" name="description" id="description" value="${task.description}">
-          </div>
-          <div>
-            <label for="dueDate">Task Due-Date</label>
-            <input type="date" name="dueDate" id="dueDate" class="border" value="${task.date}">
-          </div>
-          <div>
-            <label for="dueTime">Task Due-Time</label>
-            <input type="time" name="dueTime" id="dueTime" class="border" value="${task.time}">
-          </div>
-          <div>
-            <p>Task Priority</p>
-            <label for="urgent">Urgent</label>
-            <input type="radio" name="priority" id="urgent" class="radio border">
-            <label for="important">Important</label>
-            <input type="radio" name="priority" id="important" class="radio border">
-            <label for="lessImportant">Less Important</label>
-            <input type="radio" name="priority" id="lessImportant" class="radio border">
-          </div>
-          <div>
-            <label for="notes">Add Notes</label>
-            <input type="text" name="notes" id="notes" value="${task.notes}">
-          </div>
-          <input type="submit" value="Save Changes">
-        </form>
-        `;
-    console.log(index)
-    const parent = document.querySelector("body");
-    const formHidden = document.createElement('div');
-    formHidden.classList.add('show-form');
-    formHidden.innerHTML = formBase;
-    parent.appendChild(formHidden)
-    console.log(formHidden)
-//    formHidden.classList.remove('edit-hidden');
-    
-    
-};
 
 
-export { addTasks, appendStorage, appendProjects, completeTask };
+export { addTasks, appendStorage, appendProjects, completeTask, showTaskForm };
