+++ src/index.js
@@ -1,7 +1,7 @@
 import './style.css';
 import './Normalize.css';
 import createButGradient from './animate-dom';
-import { addTasks, appendStorage, appendProjects, completeTask } from './dom-manipulation';
+import { addTasks, appendStorage, appendProjects, completeTask, showTaskForm } from './dom-manipulation';
 
 
 createButGradient();
@@ -9,3 +9,4 @@ addTasks();
 appendStorage();
 appendProjects(true)
 completeTask();
+showTaskForm();
