+++ src/style.css
@@ -5,6 +5,7 @@ body {
 }
 
 nav {
+
 }
 .list-projects {
 
@@ -31,14 +32,14 @@ nav {
     display:none;
 }
 
-.show-form {
+/* .show-form {
     position:absolute;
     background-color:yellow;
     height:100vw;
     width: 100vw;
     top:0;
     z-index:10;
-}
+} */
 .mouse-cursor-gradient span {
   position: relative;
 }
@@ -98,20 +99,53 @@ input[type="checkbox"] {
   z-index: 3;
 }
 
-form {
+.form {
   display: flex;
   flex-flow: column wrap;
   width: 50%;
-  height: 550px;
+  height: 570px;
   justify-content: space-around;
   margin: 20px auto;
+  background-color: #fff;
+  padding: 20px;
+  border-radius: 15px;
 }
 
-form>* {
+.modal {
+  display: none;
+  z-index: 5;
+  position: fixed;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  background-color: rgba(0,0,0,0.4);
+}
+
+.close {
+  margin: 10px;
+  cursor: pointer;
+  font-size: 30px;
+  font-weight: 800;
+  color: gray;
+  float: right;
+}
+
+.close:hover,
+.close:focus {
+  color: #000;
+}
+
+.show-form {
+  display: flex;
+}
+
+.form>* {
   margin-bottom: 7px;
 }
 
-form input:not(.border) {
+.form input:not(.border) {
   display: block;
   border-color: none;
   outline: none;
@@ -142,15 +176,14 @@ form input:not(.border) {
   margin-bottom: 5px;
 }
 
-.list-item::before {
-  --width: 0;
-  content: "";
+.list-item-before {
   position: absolute;
   top: 50%;
-  width: var(--width);
+  width: 0;
   height: 5px;
   transition: width 0.5s;
-  background-color: rgba(252, 87, 101, 0.829);
+  border-radius: 20px;
+  background-color: rgba(241, 15, 109, 0.829);
 }
 
 .task-item {
